Stuck fixes including location memory and pertinent task fix
Applied fixes to rotation code
Added code to wander which spots pickups and remembers the last 10, this is to try and ensure that they had a position to go to rather than getting stuck.
Fixed selection of more pertinent tasks
Made the rotation task supplied by the manager to be an immediate action.